Policy eval, updated transition func,
Added policy evaluation.
Added get_policy_value to get the Bellman equation value.
Updated transition function according to the book.
Added helper functions:
- get_new_state_location in Game, which computes the new location given
a state and an action;
- update_policy in Predator, which takes a set of optimal moves and
updates the policy, so that only these optimal moves are considered and
the probability of the rest is 0.
- get_transformation in Predator, which given an action returns the
transformation vector (i.e. the value corresponding to the action (key)
in the actions dictionary).